Introduce HULKsStandardMessage
